extract change URL logic
