+++ assets/js/utils/filters.ts
@@ -2,6 +2,14 @@
  * External dependencies
  */
 import { getQueryArg } from '@wordpress/url';
+import { getSettingWithCoercion } from '@woocommerce/settings';
+import { isBoolean } from '@woocommerce/types';
+
+const filteringForPhpTemplate = getSettingWithCoercion(
+	'is_rendering_php_template',
+	false,
+	isBoolean
+);
 
 /**
  * Returns specified parameter from URL
@@ -17,4 +25,17 @@ export function getUrlParameter( name: string ) {
 		return null;
 	}
 	return getQueryArg( window.location.href, name );
+}
+
+/**
+ * Change the URL and reload the page if filtering for PHP templates.
+ *
+ * @param {string} newUrl New URL to be set.
+ */
+export function changeUrl( newUrl: string ) {
+	if ( filteringForPhpTemplate ) {
+		window.location.href = newUrl;
+	} else {
+		window.history.replaceState( {}, '', newUrl );
+	}
 }
